requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
django-maskurl==0.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
importlib-metadata==1.6.0
inflection==0.4.0
Markdown==3.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
gunicorn==19.6.0
django-heroku
